+++ test/Glob.Tests/GlobTests.cs
@@ -86,7 +86,7 @@ namespace GlobExpressions.Tests
         //[InlineData(@"hat\*", "hat*", "hata hatb hat hat/taco hata/taco")]
         public void TestGlobExpressions(string pattern, string positiveMatch, string negativeMatch = null)
         {
-            var glob = new Glob(pattern);
+            var glob = new Glob(pattern, GlobOptions.MatchFilenameOnly);
 
             if (positiveMatch != null)
             {
@@ -214,7 +214,7 @@ namespace GlobExpressions.Tests
         [InlineData(@"*", "a b c", "folder1/d folder2/e")]
         public void FullPathOptionTests(string pattern, string matches, string nonMatches = null)
         {
-            var glob = new Glob(pattern, GlobOptions.MatchFullPath);
+            var glob = new Glob(pattern);
             foreach (var expectedMatch in matches.Split(' '))
             {
                 Assert.True(glob.IsMatch(expectedMatch));