+++ ChangeLog.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Matching full path is now the default for Glob expressions, GlobOptions.MatchFilenameOnly
+  can be used to work as before
 ## [1.1.8]
 ### Fixed
 - Issue #59: Cannot enumerate iterator multiple times